# public/openapi.yaml
openapi: 3.1.0
info:
  title: RedRouter API
  version: 3.8.35
  description: |
    RedRouter is a local-first AI API proxy router. It provides an OpenAI-compatible
    endpoint that routes requests to multiple AI providers with load balancing,
    failover, and usage tracking.

    ## Base URLs
    - **Local**: `http://localhost:20128`

    ## Authentication
    All proxy endpoints require a Bearer token (API key managed via the dashboard).
    Management endpoints are protected when `requireLogin` is enabled.
  contact:
    name: RedRouter
  license:
    name: MIT
servers:
  - url: http://localhost:20128
    description: Local development
...
paths:
...
  /api/v1/models:
    get:
      tags: [Models]
      summary: List available models
      description: Returns all models available across configured providers.
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Model list
          content:
            application/json:
              schema:
                type: object
                properties:
                  object:
                    type: string
                    example: list
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Model"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: >
        Two bearer families are accepted. Inference API keys (typically `sk-…`)
        authorize `/v1/*`. Management routes also accept `oma_live_…` Access Tokens
        (Settings → Access Tokens / `red-router connect`) and API keys whose metadata
        includes `manage` or `admin` scope. See docs/guides/MANAGEMENT-AUTH.md.
        Bearer credentials are accepted on management routes that use this scheme;
        they are not rejected solely for being Bearer.
    ManagementSessionAuth:
      type: apiKey
      in: cookie
      name: auth_token
      description: Dashboard management session cookie (auth_token) for protected management routes. Distinct from Bearer Access Tokens and API keys. See docs/guides/MANAGEMENT-AUTH.md.
...
  schemas:
...
    Model:
      type: object
      properties:
        id:
          type: string
        object:
          type: string
          example: model
        owned_by:
          type: string
...